Splits the NBXNM setup units from the GMXForceCalculator
